Split package_ops into per-operation modules
